Cleaned up the Spare File
